+++ app/src/main/java/woowacourse/movie/view/utils/PermissionUtil.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view
+package woowacourse.movie.view.utils
 
 import android.Manifest
 import android.content.Context
